requirements.txt
tensorflow
numpy
scikit-learn
SimpleITK
nibabel
nilearn
matplotlib
pyyaml
tqdm
scikit-image
medpy